+++ src/app.tsx
@@ -21,11 +21,13 @@ import { persistor, store } from './store';
 import ThemeProvider, { toastOptions } from './theme';
 
 import '~/../assets/css/aframe.less';
+import '~/../assets/css/kbd.css';
 
 import '@fontsource/fira-sans/400.css';
 import '@fontsource/fira-sans/500.css';
 import 'react-cover-page/themes/dark.css';
 import LanguageWatcher from './i18n/LanguageWatcher';
+import HotkeyDialog from './features/hotkeys/HotkeyDialog';
 
 interface AppProps {
   readonly initialShow?: ShowLoadingRequest;
@@ -60,6 +62,7 @@ const App = ({ initialShow }: AppProps) => {
                 <AppHotkeys />
                 <MainTopLevelView />
                 <Sidebar />
+                <HotkeyDialog />
                 <Toaster toastOptions={toastOptions} />
               </>
             )}