+++ src/features/hotkeys/AppHotkeys.tsx
@@ -1,10 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import {
-  configure as configureHotkeys,
-  GlobalHotKeys,
-  type KeyMap,
-} from 'react-hotkeys';
+import { configure as configureHotkeys, GlobalHotKeys } from 'react-hotkeys';
 
 import { PLAYBACK_FPS } from '~/constants';
 import { toggleMuted } from '~/features/audio/slice';
@@ -14,14 +10,25 @@ import {
   togglePlayback,
 } from '~/features/playback/actions';
 import { switchToCameraByIndex } from '~/features/three-d/actions';
+import type { AppDispatch, RootState } from '~/store';
 
 import { keyMap } from './keymap';
-import { bindHotkeyHandlers, onlyWhenNoButtonIsFocused } from './utils';
-import type { AppDispatch, RootState } from '~/store';
+import {
+  bindHotkeyHandlers,
+  filterKeyMapByScope,
+  onlyWhenNoButtonIsFocused,
+} from './utils';
 import type { HotkeyHandler } from './types';
 import { getActiveHotkeyScope } from './selectors';
+import { showHotkeyDialog } from './slice';
 
 configureHotkeys({
+  // Needed to match '?' when it is actually 'Shift+?' (as it requires Shift
+  // to be pressed). Also needed to handle modifier keys in the keyboard
+  // handler without having to specify all possible modifier combinations in
+  // advance.
+  allowCombinationSubmatches: true,
+
   // Uncomment the next line for debugging problems with hotkeys
   // logLevel: 'debug',
 });
@@ -32,11 +39,7 @@ interface AppHotkeysProps<ScopeType = string> {
 }
 
 const AppHotkeys = ({ activeHotkeyScope, handlers }: AppHotkeysProps) => {
-  const filteredKeyMap: KeyMap = Object.fromEntries(
-    Object.entries(keyMap).filter(([, { scopes }]) =>
-      scopes.includes(activeHotkeyScope as any)
-    )
-  ) as any;
+  const filteredKeyMap = filterKeyMapByScope(keyMap, activeHotkeyScope) as any;
   return (
     <GlobalHotKeys allowChanges keyMap={filteredKeyMap} handlers={handlers} />
   );
@@ -83,6 +86,7 @@ export default connect(
         SELECT_THIRD_CAMERA: () => switchToCameraByIndex(3),
         SELECT_FOURTH_CAMERA: () => switchToCameraByIndex(4),
         SELECT_FIFTH_CAMERA: () => switchToCameraByIndex(5),
+        SHOW_HOTKEY_DIALOG: () => showHotkeyDialog(),
         TOGGLE_MUTED: () => toggleMuted(),
         TOGGLE_PLAYBACK: onlyWhenNoButtonIsFocused<AppDispatch>(togglePlayback),
       },