+++ src/features/hotkeys/selectors.ts
@@ -2,7 +2,9 @@ import { createSelector } from '@reduxjs/toolkit';
 
 import { UIMode } from '~/features/ui/modes';
 import { getCurrentMode } from '~/features/ui/selectors';
+
 import { HotkeyScope } from './keymap';
+import type { RootState } from '~/store';
 
 export const getActiveHotkeyScope = createSelector(getCurrentMode, (mode) => {
   switch (mode) {
@@ -14,3 +16,6 @@ export const getActiveHotkeyScope = createSelector(getCurrentMode, (mode) => {
       return HotkeyScope.GLOBAL;
   }
 });
+
+export const isHotkeyDialogVisible = (state: RootState) =>
+  state.hotkeys.dialogVisible;