+++ src/features/index.ts
@@ -1,6 +1,7 @@
 import { combineReducers } from 'redux';
 
 import audioReducer from './audio/slice';
+import hotkeysReducer from './hotkeys/slice';
 import playbackReducer from './playback/slice';
 import settingsReducer from './settings/slice';
 import showReducer from './show/slice';
@@ -11,6 +12,7 @@ import validationReducer from './validation/slice';
 
 export default combineReducers({
   audio: audioReducer,
+  hotkeys: hotkeysReducer,
   playback: playbackReducer,
   settings: settingsReducer,
   show: showReducer,